+++ CHANGELOG.md
@@ -8,8 +8,13 @@
 
 - Symbol references introduced
 - Context naming convention checks added: leading dot, one level depth and spaces are checking
+- Open TableResultView in separate frame with a tab context menu
 - Open In Editor action added to TableResultView. You can open content of any cell in separate editor tab
   by Alt+Enter or Alt+double left click. Can be used to open JSON, XML or any text content for quick view or edit.
+
+### Fixed
+
+- Clipboard XML exporting fixed for XML text in a cell
 
 ## [1.6.2]